new component for single cup card
